Debug print fix for action_layer when DEBUG_ACTION is defined

+++ quantum/action_layer.c
@@ -57,7 +57,9 @@ static void default_layer_state_set(layer_state_t state) {
  * Print out the hex value of the 32-bit default layer state, as well as the value of the highest bit.
  */
 void default_layer_debug(void) {
-    dprintf("%08lX(%u)", default_layer_state, get_highest_layer(default_layer_state));
+    // napa: layer_state may be uint8/16/32, depending on the configuration
+    // dprintf("%08lX(%u)", default_layer_state, get_highest_layer(default_layer_state));
+    dprintf("%08lX(%u)", (uint32_t)default_layer_state, get_highest_layer(default_layer_state));
 }
 
 /** \brief Default Layer Set
@@ -218,7 +220,9 @@ void layer_xor(layer_state_t state) {
  * Print out the hex value of the 32-bit layer state, as well as the value of the highest bit.
  */
 void layer_debug(void) {
-    dprintf("%08lX(%u)", layer_state, get_highest_layer(layer_state));
+    // napa: layer_state may be uint8/16/32, depending on the configuration
+    // dprintf("%08lX(%u)", layer_state, get_highest_layer(layer_state));
+    dprintf("%08lX(%u)", (uint32_t)layer_state, get_highest_layer(layer_state));
 }
 #endif
